requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.8.0
django-redis==5.2.0
django-simpleui==2022.4.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
